refactor: rename folder `client` into `set-wrapper`
